# src/lib/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Hindus R Us",
    "description": "An API documentation generated by swagger-autogen."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://h-dummy-app-express-app-182702284982.asia-south1.run.app/"
    }
  ],
  "paths": {
...
    "/v1/quiz/create": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "summary": "Create Quiz",
        "description": "An API route to create a new quiz",
        "responses": {
          "200": {
            "description": "Some description...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizBody"
              }
            }
          }
        }
      }
    },
    "/v1/quiz/questions/create": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "summary": "Create Quiz Questions",
        "description": "An API route to create questions required for quiz. Only one questions is allowed in each query",
        "responses": {
          "200": {
            "description": "Some description...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizQuestions"
              }
            }
          }
        }
      }
    },
    "/v1/quiz/submit": {
      "post": {
        "tags": [
          "Quiz"
        ],
        "summary": "Create Quiz Questions",
        "description": "An API route to create questions required for quiz. Only one questions is allowed in each query",
        "parameters": [
          {
            "name": "quizId",
            "in": "query",
            "description": "The quizId of quiz being submitted",
            "required": true,
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Some description...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizBody"
              }
            }
          }
        }
      }
    },
    "/v1/quiz/result": {
      "get": {
        "tags": [
          "Quiz"
        ],
        "summary": "Get Quiz Results",
        "description": "An API route to fetch quiz results",
        "parameters": [
          {
            "name": "quizId",
            "in": "query",
            "description": "The quizId of the attempted quiz",
            "required": true,
            "example": "123",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "submissionId",
            "in": "query",
            "description": "The submissionId of the attempted quiz",
            "required": true,
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Some description...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateQuizBody": {
        "type": "object",
        "properties": {
          "contentId": {
            "type": "string",
            "example": "123"
          },
          "title": {
            "type": "string",
            "example": "A Test Quiz"
          },
          "description": {
            "type": "string",
            "example": "A test description for the quiz"
          },
          "language": {
            "type": "string",
            "example": "english"
          },
          "passingScore": {
            "type": "number",
            "example": 4
          },
          "maxAttempts": {
            "type": "number",
            "example": 5
          },
          "quizType": {
            "type": "string",
            "example": "educational"
          },
          "numberOfQuestions": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "contentId",
          "title",
          "description",
          "language",
          "passingScore",
          "maxAttempts",
          "quizType",
          "numberOfQuestions"
        ]
      },
      "CreateQuizQuestions": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "A Test Quiz"
          },
          "description": {
            "type": "string",
            "example": "A test description for the quiz"
          },
          "options": {
            "type": "object",
            "properties": {
              "A": {
                "type": "string",
                "example": "Option 1"
              },
              "B": {
                "type": "string",
                "example": "Option 2"
              },
              "C": {
                "type": "string",
                "example": "Option 3"
              },
              "D": {
                "type": "string",
                "example": "Option 4"
              }
            }
          },
          "correctOption": {
            "type": "string",
            "example": "C"
          },
          "explanation": {
            "type": "string",
            "example": "Explain the answer here"
          },
          "questionType": {
            "type": "string",
            "example": "educational"
          },
          "quizType": {
            "type": "string",
            "example": "educational"
          }
        },
        "required": [
          "title",
          "description",
          "options",
          "correctOption",
          "explanation",
          "questionType",
          "quizType"
        ]
      },
      "SubmitQuizBody": {
        "type": "object",
        "properties": {
          "options": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "questionId": {
                  "type": "string",
                  "example": "123"
                },
                "selectedOption": {
                  "type": "string",
                  "example": "A"
                }
              }
            }
          }
        },
        "required": [
          "options"
        ]
      }
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}